toggle CTL is not the proper way to change status
Here's what happens when you call packetHandler.read1ByteTxRx(port, id, address):

Request: The function sends a packet out of the TX pin that
instructs the servo to read a byte of data from a specific address in its memory.

Response: The servo receives this packet, processes the request,
and then sends back a packet with the requested data.
This response is received on the RX pin.

So if we want to toggle the CTL to change direction, it needs to happen in
the packetHandler to TX out request and RX in response